Extracted style into components
