+++ src/About/About.css
@@ -1,0 +1,83 @@
+.about-content,
+.about-picture {
+  border-radius: 0.25rem;
+}
+
+.about-content {
+  position: relative;
+  padding-bottom: 0.5rem;
+
+  display: grid;
+  grid-template-rows: auto 1fr;
+}
+
+.about-content::before,
+.about-content::after {
+  position: absolute;
+  display: block;
+  transform: skew(-25deg);
+  color: var(--neutralDark);
+  font-size: 2rem;
+  font-weight: bold;
+  content: '"';
+}
+
+.about-content::before {
+  top: 0;
+  left: 0.2rem;
+}
+
+.about-content::after {
+  right: 0.5rem;
+  bottom: -1rem;
+}
+
+.about-picture {
+  grid-row: 1;
+  width: 100%;
+  max-width: 160px;
+  margin-right: 0.2rem;
+}
+
+.about-statement {
+  grid-row: 2;
+  margin-top: 0.5rem;
+  line-height: 1.25rem;
+}
+
+.about-statement p:first-child {
+  margin: 0;
+}
+
+.about-social {
+  grid-row: 3;
+}
+
+.about-social-icon {
+  height: 34px;
+  padding: 0 5px;
+  opacity: 0.33;
+  transition: all var(--animationDuration) var(--animationEasing);
+  transition-property: opacity, transform;
+}
+
+.about-social-icon:hover {
+  opacity: 0.8;
+  transform: scale(1.25);
+}
+
+@media (min-width: 640px) {
+  .about-content {
+    grid-gap: 0.8rem;
+    grid-template-columns: auto 1fr;
+  }
+
+  .about-social {
+    grid-row: 2;
+  }
+
+  .about-statement {
+    grid-row: 1 / 3;
+    margin-top: 0;
+  }
+}
