+++ src/Contact/Contact.js
@@ -1,8 +1,10 @@
+import './Contact.css';
+
 /** @jsx h */
 import { h } from 'hyperapp';
 import classy from 'classwrap';
 
-import { logger } from './logger';
+import { logger } from '../logger';
 
 const availableStyle = { color: '#27ae60' };
 
@@ -63,15 +65,15 @@ export const ContactViewModel = {
 };
 
 export const Contact = ({ state, actions }) =>
-  <section id="contact" class="contact narrow">
+  <section class="contact narrow spacer">
     <p>I'm currently <b style={availableStyle}>{availability.status}</b> for projects and consulting for <b>{availability.range}</b>.</p>
 
     <form name="form" onsubmit="return false">
-      <fieldset id="contact-fieldset" class={classy({ loading: state.pending })} disabled={state.pending || state.success}>
+      <fieldset class={classy({ loading: state.pending })} disabled={state.pending || state.success}>
         <label>
           So, how can I help?
           <textarea
-            id="contact-text"
+            class="contact-text"
             name="text"
             placeholder="Feel free to introduce yourself, describe your business idea, and how you think I could fit in the project"
             oninput={actions.setText}
@@ -84,7 +86,7 @@ export const Contact = ({ state, actions }) =>
         <label>
           And how can I reach you?
           <input
-            id="contact-email"
+            class="contact-email"
             type="email"
             name="email"
             placeholder="Email"
@@ -99,12 +101,12 @@ export const Contact = ({ state, actions }) =>
       </fieldset>
     </form>
 
-    <div id="contact-success" class={classy({ shown: state.success })}>
+    <div class={classy(['contact-success', { shown: state.success }])}>
       Thanks for reaching out 😊 I'll get back to you soon!
     </div>
 
-    <div id="contact-error" class={classy({ shown: state.error })}>
-      Oops! Something went wrong 😐 The error has been logged, I'll see what I can do about it. And don't worry - you can still reach me at&nbsp;
+    <div class={classy(['contact-error', { shown: state.error }])}>
+      Oops! Something went wrong 😐 Sorry about that, the error has been logged, I'll see what I can do about it. And don't worry - you can still reach me at&nbsp;
       {/* NOTE: Email text is used for mailto body: http://www.angelfire.com/dc/html-webmaster/mailto.htm */}
       <a target="_blank" href={`mailto:${myEmailAddress}?body=${state.text}`}>{myEmailAddress}</a>
     </div>
