+++ src/Footer/Footer.js
@@ -1,3 +1,5 @@
+import './Footer.css';
+
 /** @jsx h */
 import { app, h } from 'hyperapp';
 
@@ -5,7 +7,8 @@ const thisYear = new Date().getFullYear();
 
 export const Footer = () => [
   <div class="footer-stripe"></div>,
-  <footer id="footer">
+
+  <footer class="footer">
     <div class="footer-content narrow">
       <p>&copy; {thisYear} Vladimir Sabev</p>
       <p>Illustration by <a target="_blank" href="https://twitter.com/spovv">@spovv</a></p>
