+++ src/Profile/Profile.js
@@ -1,10 +1,12 @@
-import './assets/avatar.svg';
+import './avatar.svg';
+
+import './Profile.css';
 
 /** @jsx h */
 import { app, h } from 'hyperapp';
 
 export const Profile = () =>
-  <section id="profile" class="profile narrow">
+  <section class="profile narrow">
     <img class="profile-avatar" src="avatar.svg" alt="Vladimir Sabev's Avatar" />
     <div class="profile-text">
       <h1 class="profile-title">Vladimir Sabev</h1>
