+++ src/index.ejs
@@ -13,7 +13,7 @@
     <meta content="on" http-equiv="cleartype">
 
     <base href="/">
-    <link rel="shortcut icon" href="avatar.png">
+    <link rel="shortcut icon" href="favicon.png">
 
     <% const values = (obj) => Object.keys(obj).map((key) => obj[key]) %>
     <% const fonts = htmlWebpackPlugin.options.style.fonts; %>
