+++ src/index.js
@@ -1,4 +1,5 @@
-import './assets/avatar.png';
+import './favicon.png';
+
 import './animation.css';
 import './style.css';
