+++ src/style.css
@@ -102,216 +102,6 @@ button {
   padding: 1.5rem;
 }
 
-.about,
-.contact,
-.post-list {
+.spacer {
   margin-bottom: 3rem;
-}
-
-/* Profile */
-.profile {
-  margin-top: 3rem;
-  margin-bottom: 3rem;
-  font-family: var(--fontProfile), sans-serif;
-  text-align: center;
-}
-
-/* NOTE: I can't quite decide whether it looks better with or without the grid */
-/* @media (min-width: 800px) {
-  .profile {
-    display: grid;
-    grid-auto-flow: column;
-    justify-content: center;
-    align-items: center;
-  }
-} */
-
-.profile-avatar {
-  width: 180px;
-  margin: 1.5rem;
-}
-
-.profile-title,
-.profile-subtitle {
-  -webkit-margin-before: 0;
-  -webkit-margin-after: 0;
-}
-
-.profile-title {
-  font-size: 3rem;
-  font-weight: normal;
-  margin: 0;
-}
-
-.profile-subtitle {
-  font-size: 1.5rem;
-}
-
-/* About */
-.about-content,
-.about-picture {
-  border-radius: 0.25rem;
-}
-
-.about-content {
-  position: relative;
-  padding-bottom: 0.5rem;
-
-  display: grid;
-  grid-template-rows: auto 1fr;
-}
-
-.about-content::before,
-.about-content::after {
-  position: absolute;
-  display: block;
-  transform: skew(-25deg);
-  color: var(--neutralDark);
-  font-size: 2rem;
-  font-weight: bold;
-  content: '"';
-}
-
-.about-content::before {
-  top: 0;
-  left: 0.2rem;
-}
-
-.about-content::after {
-  right: 0.5rem;
-  bottom: -1rem;
-}
-
-.about-picture {
-  grid-row: 1;
-  width: 100%;
-  max-width: 160px;
-  margin-right: 0.2rem;
-}
-
-.about-statement {
-  grid-row: 2;
-  margin-top: 0.5rem;
-  line-height: 1.25rem;
-}
-
-.about-statement p:first-child {
-  margin: 0;
-}
-
-.about-social {
-  grid-row: 3;
-}
-
-.about-social-icon {
-  height: 34px;
-  padding: 0 5px;
-  opacity: 0.33;
-  transition: all var(--animationDuration) var(--animationEasing);
-  transition-property: opacity, transform;
-}
-
-.about-social-icon:hover {
-  opacity: 0.8;
-  transform: scale(1.25);
-}
-
-@media (min-width: 640px) {
-  .about-content {
-    grid-gap: 0.8rem;
-    grid-template-columns: auto 1fr;
-  }
-
-  .about-social {
-    grid-row: 2;
-  }
-
-  .about-statement {
-    grid-row: 1 / 3;
-    margin-top: 0;
-  }
-}
-
-/* Contact */
-#contact-text,
-#contact-email {
-  width: 100%;
-  margin-top: 0.25rem;
-}
-
-.contact-submit {
-  width: 100%;
-  background: var(--success);
-  color: var(--neutralLighter);
-  font-size: 1.5rem;
-  font-variant: small-caps;
-}
-
-#contact-success,
-#contact-error {
-  display: none;
-  opacity: 0;
-  transition: opacity var(--animationDuration) var(--animationEasing);
-  font-size: 1.5rem;
-}
-
-#contact-success {
-  color: var(--success);
-}
-
-#contact-error {
-  color: var(--danger);
-}
-
-#contact-success.shown,
-#contact-error.shown {
-  display: block;
-  opacity: 1;
-}
-
-/* Posts */
-.post-item {
-  margin-bottom: 2rem;
-}
-
-.post-item-title {
-  font-size: 22px;
-  color: var(--foregroundDark);
-  transition: color var(--animationDuration) var(--animationEasing);
-}
-
-.post-item-title:hover {
-  color: var(--foregroundLight);
-}
-
-.post-item-text {
-  margin: 0.33rem 0;
-}
-
-.post-item-date {
-  font-size: 14px;
-  color: var(--foreground);
-}
-
-/* Footer */
-.footer-stripe {
-  height: 5px;
-  background: var(--info);
-  background: linear-gradient(to right, var(--info), var(--success));
-}
-
-footer {
-  background: var(--primary);
-  padding: 1rem;
-  color: var(--neutralLighter);
-}
-
-.footer-content {
-  display: grid;
-  grid-template-columns: 1fr auto;
-}
-
-.footer-content a {
-  color: var(--neutralLighter);
-  border-bottom: 1px dotted var(--neutralLight);
 }
